Added QuickLook generated previews (works with many different file formats) on OS X using qlpreview. Currently must be executed as an offline process by a 'real' user otherwise qlpreview returns a generic document item. Hope to resolve that shortly.
Removed 'rating' field from default field set (I don't deem it generally useful to have an admin rating and user ratings are probably preferable, of course this can be changed per-system anyway)

Made writing metadump.xml files the default, because disk usage is minor and it is a useful backup.

Altered create_previews.php so it can work in a single process with the process forking functions missing (as with the default PHP install on OS X Snow Leopard).